requirements.txt
python-dotenv>=1.0.0
openai>=1.6.1
aiohttp>=3.9.1
async-timeout>=4.0.3